requirements.txt
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
cryptography==39.0.1
cupy==10.1.0
cytoolz==0.12.0
dask==2023.4.1
einops==0.6.1
fastrlock==0.6
filelock==3.9.0
fsspec==2023.4.0
h5py==3.7.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
lazy_loader==0.1
locket==1.0.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.8.4
numpy==1.24.3
packaging==23.0
partd==1.2.0
Pillow==9.5.0
pip==23.0.1
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
scikit-image==0.20.0
scipy==1.8.1
setuptools==66.0.0
six==1.16.0
thop==0.1.1.post2209072238
tifffile==2020.10.1
timm==0.6.13
toolz==0.12.0
torch==1.8.2+cu102
torchaudio==0.8.2
torchvision==0.9.2+cu102
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
wheel==0.38.4
win-inet-pton==1.1.0
zipp==3.11.0
